+++ tests/Normalizer/ObjectNormalizer/ValueWithContextNormalizerTest.php
@@ -1,0 +1,49 @@
+<?php declare(strict_types=1);
+
+namespace Tests\Torr\SimpleNormalizer\Normalizer\ObjectNormalizer;
+
+use PHPUnit\Framework\TestCase;
+use Symfony\Component\DependencyInjection\ServiceLocator;
+use Tests\Torr\SimpleNormalizer\Fixture\DummyVO;
+use Torr\SimpleNormalizer\Data\ValueWithContext;
+use Torr\SimpleNormalizer\Normalizer\ObjectNormalizer\ValueWithContextNormalizer;
+use Torr\SimpleNormalizer\Normalizer\SimpleNormalizer;
+use Torr\SimpleNormalizer\Normalizer\SimpleObjectNormalizerInterface;
+
+final class ValueWithContextNormalizerTest extends TestCase
+{
+	/**
+	 *
+	 */
+	public function testContextPassing () : void
+	{
+		$value = new ValueWithContext(
+			new DummyVO(5),
+			[
+				"o" => "hai",
+			],
+		);
+
+		$dummyNormalizer = $this->createMock(SimpleObjectNormalizerInterface::class);
+		$dummyNormalizer
+			->expects(self::once())
+			->method("normalize")
+			->with(
+				$value->value,
+				[
+					"test" => 123,
+					"o" => "hai",
+				],
+			);
+
+		$normalizer = new SimpleNormalizer(new ServiceLocator([
+			ValueWithContext::class => fn () => new ValueWithContextNormalizer(),
+			DummyVO::class => fn () => $dummyNormalizer,
+		]));
+
+		$normalizer->normalize($value, [
+			"test" => 123,
+			"o" => 5,
+		]);
+	}
+}